add scenario chaincode calling chaincode

+++ blockchain-explorer/stop.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
-docker rm -fv $(docker ps -a|grep blockchain-explorer|awk '{print $1}')
+CONTAINER=$(docker ps -a|grep blockchain-explorer|awk '{print $1}')
+if [[ ${#CONTAINER} -gt 0 ]]; then
+    docker rm -fv $(docker ps -a|grep blockchain-explorer|awk '{print $1}')
+fi
